separated socket and gameState from app.js into there own files
